+++ Changes.md
@@ -7,6 +7,7 @@ refer to [https://github.com/msimerson/Mail-Toaster-6/commits/master](https://gi
 
 - config: mv config into ./conf.d/
 - config: split service portions into $service.conf files
+- config: tighten up perms on .mt6 files
 - mysql: drop vestigial SQUIRREL_SQL/ROUNDCUBE_SQL guard
 - config: move jail-private defaults into their provision script
 - spamassassin: drop redundant MAXMIND_LICENSE_KEY check, RELAY_COUNTRY always builds